renamed things, lots of known lint errors (just testing for now)
